Separated the utilty files into their own project and now reference them in the main project.
