+++ ProjectStone.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31515.178
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProjectStone", "ProjectStone\ProjectStone.csproj", "{B9A27909-9EC8-4F23-92B8-C43276B2C2C6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProjectStone", "ProjectStone\ProjectStone.csproj", "{B9A27909-9EC8-4F23-92B8-C43276B2C2C6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProjectStone_Utility", "ProjectStone_Utility\ProjectStone_Utility.csproj", "{69229B0D-DC57-461B-BA10-281C3F46FF19}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{B9A27909-9EC8-4F23-92B8-C43276B2C2C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B9A27909-9EC8-4F23-92B8-C43276B2C2C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B9A27909-9EC8-4F23-92B8-C43276B2C2C6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{69229B0D-DC57-461B-BA10-281C3F46FF19}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{69229B0D-DC57-461B-BA10-281C3F46FF19}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{69229B0D-DC57-461B-BA10-281C3F46FF19}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{69229B0D-DC57-461B-BA10-281C3F46FF19}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
