+++ ProjectStone/Areas/Identity/Pages/Account/Register.cshtml
@@ -1,7 +1,7 @@
 ﻿@page
 @model RegisterModel
 
-@if (User.IsInRole(ProjectStone.WebConstants.AdminRole))
+@if (User.IsInRole(ProjectStone_Utility.WebConstants.AdminRole))
 {
     <h1>Register an Admin Account</h1>
 }
@@ -39,7 +39,7 @@ else
                 <input asp-for="Input.ConfirmPassword" class="form-control"/>
                 <span asp-validation-for="Input.ConfirmPassword" class="text-danger"></span>
             </div>
-            @if (User.IsInRole(ProjectStone.WebConstants.AdminRole))
+            @if (User.IsInRole(ProjectStone_Utility.WebConstants.AdminRole))
             {
                 <button type="submit" class="btn btn-warning">Register Admin User</button>
             }
