+++ ProjectStone/Areas/Identity/Pages/Account/Register.cshtml.cs
@@ -14,6 +14,7 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.Logging;
 using ProjectStone.Models;
+using ProjectStone_Utility;
 
 namespace ProjectStone.Areas.Identity.Pages.Account
 {
