+++ ProjectStone/Views/Cart/Index.cshtml
@@ -1,4 +1,5 @@
-﻿@model IEnumerable<Product>
+﻿@using ProjectStone_Utility
+@model IEnumerable<Product>
 
 @if (Model.Any())
 {
@@ -38,7 +39,7 @@
                             {
                                 <div class="row h-100">
                                     <div class="col-4 col-md-2 text-center py-2">
-                                        <img src="@WebConstants.ImagePath@item.Image" class="rounded" width="100%" alt="Product Image"/>
+                                        <img src="@WebConstants.ImagePath@item.Image" class="rounded" style="width: 100%;" alt="Product Image"/>
                                     </div>
                                     <div class="col-8 col-lg-7">
                                         <h5>@item.Name</h5>
