+++ ProjectStone/Views/Home/Details.cshtml
@@ -1,4 +1,5 @@
-﻿@model ProjectStone.Models.ViewModels.DetailsViewModel
+﻿@using ProjectStone_Utility
+@model ProjectStone.Models.ViewModels.DetailsViewModel
 
 <div class="">
     <form method="post">
@@ -17,7 +18,7 @@
                     <div class="container rounded p-2">
                         <div class="row">
                             <div class="col-12 col-lg-4 p-1 text-center">
-                                <img src="@WebConstants.ImagePath@Model.Product.Image" width="100%" class="rounded" alt="Product Image"/>
+                                <img src="@WebConstants.ImagePath@Model.Product.Image" style="width: 100%;" class="rounded" alt="Product Image"/>
                             </div>
                             <div class="col-12 col-lg-8">
                                 <div class="row pl-3">
