+++ ProjectStone/Views/Product/Delete.cshtml
@@ -1,4 +1,5 @@
-﻿@model ProjectStone.Models.Product
+﻿@using ProjectStone_Utility
+@model ProjectStone.Models.Product
 
 <form method="post">
   @* Grab the Id when we post it *@
