+++ ProjectStone/Views/Product/Upsert.cshtml
@@ -1,4 +1,5 @@
-﻿@model ProjectStone.Models.ViewModels.ProductViewModel
+﻿@using ProjectStone_Utility
+@model ProjectStone.Models.ViewModels.ProductViewModel
 
 @{
     var title = "Create Product";
