+++ ProjectStone/Views/Shared/_IndividualProductCard.cshtml
@@ -1,4 +1,5 @@
-﻿@model ProjectStone.Models.Product
+﻿@using ProjectStone_Utility
+@model ProjectStone.Models.Product
 
 <div class="col-lg-4 col-md-6 pb-4 filter @Model.Category.Name.Replace(' ','_')">
     <!-- Card-->
