+++ ProjectStone/Views/Shared/_Layout.cshtml
@@ -2,7 +2,7 @@
 @using Microsoft.AspNetCore.Http
 @inject IHttpContextAccessor _httpContextAccessor
 @* Add Utilities as well *@
-@using ProjectStone.Utility
+@using ProjectStone_Utility
 
 <!DOCTYPE html>
 <html lang="en">
